+++ backend/src/main/java/moaon/backend/article/domain/ArticleDocument.java
@@ -1,4 +1,4 @@
-package moaon.backend.es;
+package moaon.backend.article.domain;
 
 import static java.util.stream.Collectors.toSet;
 
@@ -11,9 +11,6 @@ import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
-import moaon.backend.article.domain.Article;
-import moaon.backend.article.domain.Sector;
-import moaon.backend.article.domain.Topic;
 import moaon.backend.techStack.domain.TechStack;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.elasticsearch.annotations.DateFormat;
