+++ backend/src/main/java/moaon/backend/article/repository/es/ArticleDocumentRepository.java
@@ -1,5 +1,7 @@
-package moaon.backend.es;
+package moaon.backend.article.repository.es;
 
+import moaon.backend.article.domain.ArticleDocument;
+import moaon.backend.article.dto.ArticleESQuery;
 import org.springframework.data.elasticsearch.core.SearchHits;
 
 public interface ArticleDocumentRepository {
