+++ backend/src/test/java/moaon/backend/api/article/ArticleESApiTest.java
@@ -10,8 +10,8 @@ import moaon.backend.article.domain.Topic;
 import moaon.backend.article.dto.ArticleContent;
 import moaon.backend.article.dto.ArticleResponse;
 import moaon.backend.article.repository.ArticleRepository;
-import moaon.backend.es.ArticleDocument;
-import moaon.backend.es.ArticleDocumentRepository;
+import moaon.backend.article.domain.ArticleDocument;
+import moaon.backend.article.repository.es.ArticleDocumentRepository;
 import moaon.backend.fixture.ArticleFixtureBuilder;
 import moaon.backend.fixture.Fixture;
 import moaon.backend.project.domain.Project;
